+++ src/dll/Hooks/ValidateScripts.cpp
@@ -4,6 +4,7 @@
 #include "Hook.hpp"
 #include "Systems/ScriptCompilationSystem.hpp"
 #include "RED4ext/Scripting/ScriptReport.hpp"
+#include "ScriptValidationError.hpp"
 
 namespace
 {
@@ -87,113 +88,4 @@ bool Hooks::ValidateScripts::Detach()
 
     isAttached = result != NO_ERROR;
     return !isAttached;
-}
-
-ValidationError ValidationError::FromString(const char* str)
-{
-    ValidationErrorType type = ValidationErrorType::Unknown;
-    char name[64] = {0};
-    char parent[64] = {0};
-
-    if (sscanf_s(str, "Missing native class '%[^']'", name, (int)sizeof(name)) == 1)
-    {
-        type = ValidationErrorType::MissingClass;
-    }
-    else if (sscanf_s(str, "Missing native global function '%[^']'", name, (int)sizeof(name)) == 1)
-    {
-        type = ValidationErrorType::MissingGlobalFunction;
-    }
-    else if (sscanf_s(str, "Missing native function '%[^']' in native class '%[^']'", name, (int)sizeof(name), parent,
-                      (int)sizeof(parent)) == 2)
-    {
-        type = ValidationErrorType::MissingMethod;
-    }
-    else if (sscanf_s(str, "Missing native property '%[^']' in native class '%[^']'", name, (int)sizeof(name), parent,
-                      (int)sizeof(parent)) == 2)
-    {
-        type = ValidationErrorType::MissingProperty;
-    }
-    else if (sscanf_s(str, "Missing base class '%[^']' of native class '%[^']'", parent, (int)sizeof(parent), name,
-                      (int)sizeof(name)) == 2)
-    {
-        type = ValidationErrorType::MissingBaseClass;
-    }
-    else if (sscanf_s(
-                 str,
-                 "Native class '%[^']' has declared base class '%[^']' that is different than current one '%*[^']'",
-                 name, (int)sizeof(name), parent, (int)sizeof(parent)) == 2)
-    {
-        type = ValidationErrorType::BaseClassMismatch;
-    }
-    else if (sscanf_s(str, "Imported property '%[^.].%[^']' type '%*[^']' does not match with the native one '%*[^']'",
-                      parent, (int)sizeof(parent), name, (int)sizeof(name)) == 2)
-    {
-        type = ValidationErrorType::PropertyTypeMismatch;
-    }
-
-    return { .type = type, .name = name, .parent = parent };
-}
-
-std::optional<SourceRef> ValidationError::GetSourceRef() const
-{
-    auto& sourceRepo = App::Get()->GetScriptCompilationSystem()->GetSourceRefRepository();
-
-    try
-    {
-        switch (type)
-        {
-        case ValidationErrorType::MissingClass:
-            return sourceRepo.GetClass(name);
-        case ValidationErrorType::MissingGlobalFunction:
-            return sourceRepo.GetFunction(name);
-        case ValidationErrorType::MissingMethod:
-            return sourceRepo.GetMethod(name, parent);
-        case ValidationErrorType::MissingProperty:
-            return sourceRepo.GetProperty(name, parent);
-        case ValidationErrorType::MissingBaseClass:
-            return sourceRepo.GetClass(name);
-        case ValidationErrorType::BaseClassMismatch:
-            return sourceRepo.GetClass(name);
-        case ValidationErrorType::PropertyTypeMismatch:
-            return sourceRepo.GetProperty(name, parent);
-        default:
-            return {};
-        }
-    }
-    catch (std::out_of_range)
-    {
-        return {};
-    }
-}
-
-std::string WritePopupMessage(const std::vector<ValidationError>& errors)
-{
-    std::unordered_set<std::string_view> faultyFiles;
-
-    for (const auto& error : errors)
-    {
-        auto ref = error.GetSourceRef();
-        if (ref)
-        {
-            faultyFiles.insert(ref->file);
-        }
-    }
-
-    if (faultyFiles.empty())
-    {
-        return "";
-    }
-
-    std::string message =
-        "The scripts below contain invalid native definitions and will prevent your game from starting:\n";
-    for (const auto& file : faultyFiles)
-    {
-        fmt::format_to(std::back_inserter(message), "- {}\n", file);
-    }
-    fmt::format_to(std::back_inserter(message),
-                   "\n"
-                   "Check for updates of the mods that added these files and if you still see this "
-                   "message after updating them, they'll have to be uninstalled.\n"
-                   "More details about the errors can be found in the logs.\n");
-    return message;
 }
